+++ src/game/index.ts
@@ -1,0 +1,15 @@
+import GameArea from './gamearea/gamearea';
+import Renderer from './gamearea/renderer';
+
+import Console from './sidebar/console';
+import MatchQueue from './sidebar/matchqueue';
+import MatchRunner from './sidebar/matchrunner';
+import Stats from './sidebar/stats';
+
+import TickCounter from './fps';
+import {NextStepSchema} from './nextstep';
+import NextStep from './nextstep';
+
+export {GameArea, Renderer};
+export {Console, MatchQueue, MatchRunner, Stats};
+export {TickCounter, NextStepSchema, NextStep};
