+++ src/mapeditor/forms/archon.ts
@@ -1,9 +1,8 @@
 import * as cst from '../../constants';
 
-import {UnitForm} from './unitform';
-
-import {MapUnit} from '../renderer';
 import Victor = require('victor');
+
+import {UnitForm, MapUnit} from '../index';
 
 export default class ArchonForm implements UnitForm {
