+++ src/static/css/style.css
@@ -56,7 +56,7 @@
 }
 
 #mapEditor {
-  font-family: "Tahoma, sans serif";
+  font-family: Tahoma, sans-serif;
   font-size: 14px;
 }
